feat(cli): consolidate formatter and linter options into eslint/prettier; remove redundant prettier installer
